requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt --resolver=backtracking requirements-dev.in requirements.txt
#
aiobotocore==2.5.0
    # via
    #   -r requirements.txt
    #   s3fs
aiohttp==3.8.4
    # via
    #   -r requirements.txt
    #   aiobotocore
    #   s3fs
aioitertools==0.11.0
    # via
    #   -r requirements.txt
    #   aiobotocore
aiosignal==1.3.1
    # via
    #   -r requirements.txt
    #   aiohttp
async-timeout==4.0.2
    # via
    #   -r requirements.txt
    #   aiohttp
attrs==23.1.0
    # via
    #   -r requirements.txt
    #   aiohttp
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.185
    # via
    #   -r requirements.txt
    #   aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via
    #   -r requirements.txt
    #   aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.155
    # via
    #   -r requirements.txt
    #   aws-cdk-lib
aws-cdk-lib==2.82.0
    # via -r requirements.txt
botocore==1.29.76
    # via
    #   -r requirements.txt
    #   aiobotocore
build==0.10.0
    # via pip-tools
cattrs==22.2.0
    # via
    #   -r requirements.txt
    #   jsii
charset-normalizer==3.1.0
    # via
    #   -r requirements.txt
    #   aiohttp
click==8.1.3
    # via pip-tools
constructs==10.2.40
    # via
    #   -r requirements.txt
    #   aws-cdk-lib
exceptiongroup==1.1.1
    # via
    #   -r requirements.txt
    #   cattrs
    #   pytest
frozenlist==1.3.3
    # via
    #   -r requirements.txt
    #   aiohttp
    #   aiosignal
fsspec==2023.5.0
    # via
    #   -r requirements.txt
    #   s3fs
idna==3.4
    # via
    #   -r requirements.txt
    #   yarl
importlib-resources==5.12.0
    # via
    #   -r requirements.txt
    #   jsii
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   -r requirements.txt
    #   botocore
jsii==1.82.0
    # via
    #   -r requirements.txt
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
multidict==6.0.4
    # via
    #   -r requirements.txt
    #   aiohttp
    #   yarl
mypy==1.3.0
    # via -r requirements-dev.in
mypy-extensions==1.0.0
    # via mypy
numpy==1.24.3
    # via
    #   -r requirements.txt
    #   pandas
    #   pyarrow
packaging==23.1
    # via
    #   build
    #   pytest
pandas==2.0.1
    # via -r requirements.txt
pip-tools==6.13.0
    # via -r requirements-dev.in
pluggy==1.0.0
    # via pytest
publication==0.0.3
    # via
    #   -r requirements.txt
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
pyarrow==12.0.0
    # via -r requirements.txt
pyproject-hooks==1.0.0
    # via build
pytest==7.3.1
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via
    #   -r requirements.txt
    #   botocore
    #   jsii
    #   pandas
pytz==2023.3
    # via
    #   -r requirements.txt
    #   pandas
s3fs==2023.5.0
    # via -r requirements.txt
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
tomli==2.0.1
    # via
    #   build
    #   mypy
    #   pyproject-hooks
    #   pytest
typeguard==2.13.3
    # via
    #   -r requirements.txt
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.6.3
    # via
    #   -r requirements.txt
    #   jsii
    #   mypy
tzdata==2023.3
    # via
    #   -r requirements.txt
    #   pandas
urllib3==1.26.16
    # via
    #   -r requirements.txt
    #   botocore
wheel==0.40.0
    # via pip-tools
wrapt==1.15.0
    # via
    #   -r requirements.txt
    #   aiobotocore
yarl==1.9.2
    # via
    #   -r requirements.txt
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
